requirements.txt
absl-py==2.3.1
altgraph==0.17.4
argparse-addons==0.12.0
asgiref==3.6.0
asttokens==3.0.0
astunparse==1.6.3
beautifulsoup4==4.12.3
bincopy==20.0.0
blinker==1.9.0
Brotli==1.1.0
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.0.7
cryptography==43.0.0
cycler==0.11.0
DateTime==5.1
debugpy==1.8.14
decorator==5.2.1
distlib==0.3.6
Django==4.2.1
django-ckeditor==6.7.2
django-extended-choices==1.3.3
django-js-asset==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
executing==2.2.0
filelock==3.12.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.39.3
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
humanfriendly==10.0
idna==3.7
imageio==2.34.1
imbalanced-learn==0.13.0
imblearn==0.0
inflate64==1.0.0
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isodate==0.6.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.19.2
Jinja2==3.1.4
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.10.0
keyring==25.2.1
kiwisolver==1.4.4
ldap3==2.9.1
libclang==18.1.1
lxml==5.2.2
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
more-itertools==10.2.0
multivolumefile==0.2.3
Naked==0.1.32
namex==0.1.0
nest-asyncio==1.6.0
numpy==1.26.4
opencv-python==4.11.0.86
openpyxl==3.1.2
opt_einsum==3.4.0
optree==0.16.0
packaging==23.1
pandas==2.0.1
parso==0.8.4
pefile==2023.2.7
Pillow==9.5.0
Pint==0.22
pipenv==2023.4.29
platformdirs==3.5.0
prompt_toolkit==3.0.51
protobuf==5.29.5
psutil==6.0.0
psycopg2==2.9.9
pure_eval==0.2.3
py7zr==0.22.0
pyasn1==0.6.1
pybcj==1.0.2
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pyelftools==0.31
pygame==2.5.2
Pygments==2.19.2
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.5
pyparsing==3.0.9
pyppmd==1.1.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyreadline3==3.4.1
pytesseract==0.3.13
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyunpack==0.3
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==27.0.0
pyzstd==0.16.1
rdflib==7.0.0
regex==2023.5.5
requests==2.31.0
rich==14.0.0
scikit-learn==1.6.1
scipy==1.16.0
seaborn==0.13.2
sequential==1.0.0
setuptools==69.5.1
shellescape==3.8.1
six==1.16.0
sklearn-compat==0.1.3
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
texttable==1.7.0
threadpoolctl==3.6.0
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.6.2
tzdata==2023.3
Unidecode==1.3.8
urllib3==2.2.1
virtualenv==20.23.0
virtualenv-clone==0.5.7
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.43.0
wrapt==1.17.2
xlrd==2.0.1
xlwings==0.30.7
xmltodict==0.13.0
zope.interface==6.0
